src/requirements.txt
pydantic==1.8.2
jsonpatch==1.32
fastapi==0.75.0
pyOpenSSL==22.0.0
